+++ practice_with_YumemiWeather/Data/WeatherState.swift
@@ -6,9 +6,34 @@
 //
 
 import Foundation
+import UIKit
 
 enum WeatherState: String {
     case sunny = "sunny"
     case rainy = "rainy"
     case cloudy = "cloudy"
+}
+
+extension WeatherState {
+    var image: UIImage? {
+        switch self {
+        case .sunny:
+            return UIImage(named: "sunny")
+        case .cloudy:
+            return UIImage(named: "cloudy")
+        case .rainy:
+            return UIImage(named: "rainy")
+        }
+    }
+    
+    var color: UIColor {
+        switch self {
+        case .sunny:
+            return .orange
+        case .cloudy:
+            return .gray
+        case .rainy:
+            return .blue
+        }
+    }
 }
